#full entries module documentation
