+++ entries/src/main/kotlin/com/source/bricks/entry/Entry.kt
@@ -1,3 +1,13 @@
 package com.source.bricks.entry
 
-data class Entry<T>(val version: Long = 0, val data: T)
+
+/**
+ * Defines entries used in [com.source.bricks.entry.scope.EntryScope]
+ *
+ * @param version entry version (migrations)
+ * @param data entry value
+ * */
+data class Entry<T>(
+    val version: Long = 0,
+    val data: T
+)
