+++ entries/src/main/kotlin/com/source/bricks/entry/EntryProperty.kt
@@ -2,6 +2,13 @@ package com.source.bricks.entry
 
 import kotlin.properties.ReadWriteProperty
 
+
+/**
+ * Entry property delegation description
+ * */
 interface EntryProperty<T> : ReadWriteProperty<Any?, T> {
+    /**
+     * Removes entry from scope
+     * */
     fun remove()
 }
