+++ entries/src/main/kotlin/com/source/bricks/entry/transcoder/EntryEncoder.kt
@@ -1,7 +1,11 @@
 package com.source.bricks.entry.transcoder
 
-import com.source.bricks.reflect.tClass
+import com.source.bricks.reflect.TClass
 
+
+/**
+ * Encodes input object
+ * */
 interface EntryEncoder<Raw> {
     fun <Type> encode(input: Type, cls: TClass<Type>): Raw?
 }
