+++ entries/src/main/kotlin/com/source/bricks/entry/transcoder/EntryTranscoder.kt
@@ -1,5 +1,9 @@
 package com.source.bricks.entry.transcoder
 
+
+/**
+ * Defines Encoder and Decoder combined together
+ * */
 interface EntryTranscoder<Raw> :
     EntryEncoder<Raw>,
-        EntryDecoder<Raw>
+    EntryDecoder<Raw>
